Lab_10/requirements.txt
pygame>=2.5
PyOpenGL>=3.1.7
PyOpenGL-accelerate>=3.1.7
PyGLM>=2.7
numpy>=1.24

